Fixed typo and compilation warnings/error under MSVC when building for x64. (Jonathan Potter, GP Software)
git-svn-id: svn://dev.exiv2.org/svn@1895 b7c8b350-86e7-0310-a4b4-de8f6a8f16a3

+++ branches/unstable/src/pngimage.cpp
@@ -287,7 +287,7 @@ namespace Exiv2 {
                 {
                     // Update Comment data to a new PNG chunk
                     std::string chunk = PngChunk::makeMetadataChunk(comment_, mdComment);
-                    if (outIo.write((const byte*)chunk.data(), chunk.size()) != (long)chunk.size())
+                    if (outIo.write((const byte*)chunk.data(), static_cast<long>(chunk.size())) != (long)chunk.size())
                     {
                         throw Error(21);
                     }
@@ -304,7 +304,7 @@ namespace Exiv2 {
                         std::string rawExif =   std::string(exifHeader, 6)
                                               + std::string((const char*)&blob[0], blob.size());
                         std::string chunk = PngChunk::makeMetadataChunk(rawExif, mdExif);
-                        if (outIo.write((const byte*)chunk.data(), chunk.size()) != (long)chunk.size())
+                        if (outIo.write((const byte*)chunk.data(), static_cast<long>(chunk.size())) != (long)chunk.size())
                         {
                             throw Error(21);
                         }
@@ -319,7 +319,7 @@ namespace Exiv2 {
                     {
                         std::string rawIptc((const char*)newPsData.pData_, newPsData.size_);
                         std::string chunk = PngChunk::makeMetadataChunk(rawIptc, mdIptc);
-                        if (outIo.write((const byte*)chunk.data(), chunk.size()) != (long)chunk.size())
+                        if (outIo.write((const byte*)chunk.data(), static_cast<long>(chunk.size())) != (long)chunk.size())
                         {
                             throw Error(21);
                         }
@@ -336,7 +336,7 @@ namespace Exiv2 {
                 if (xmpPacket_.size() > 0) {
                     // Update XMP data to a new PNG chunk
                     std::string chunk = PngChunk::makeMetadataChunk(xmpPacket_, mdXmp);
-                    if (outIo.write((const byte*)chunk.data(), chunk.size()) != (long)chunk.size()) {
+                    if (outIo.write((const byte*)chunk.data(), static_cast<long>(chunk.size())) != (long)chunk.size()) {
                         throw Error(21);
                     }
                 }
